🔨[refactor] 워크스페이스 초대 코드 제거 #109
