--- src/main/java/com/daon/backend/task/dto/workspace/InviteMemberRequestDto.java
@@ -1,13 +1,0 @@
-package com.daon.backend.task.dto.workspace;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class InviteMemberRequestDto {
-
-    private String email;
-}
