--- src/main/java/com/daon/backend/task/service/DbMemberProvider.java
@@ -1,6 +1,0 @@
-package com.daon.backend.task.service;
-
-public interface DbMemberProvider {
-
-    String getMemberIdByEmail(String email);
-}
