+++ src/main/resources/sql/data.sql
@@ -15,15 +15,15 @@ values ('2023-11-11 17:38:13', '2023-11-11 17:38:13', 'test1@email.com', '78cfb9
        ('2023-11-11 17:38:13', '2023-11-11 17:38:13', 'test8@email.com', 'c152ed2e-1978-4100-bf26-b0014af642b4');
 
 
-INSERT INTO workspace (created_at, modified_at, description, division, image_url, join_code, subject, title)
-VALUES ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, null, '개인용도', 'USER1 개인워크스페이스'),
-       ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, null, '개인용도', 'USER2 개인워크스페이스'),
-       ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, null, '개인용도', 'USER3 개인워크스페이스'),
-       ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, null, '개인용도', 'USER4 개인워크스페이스'),
-       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, null, '그룹 테스트', '그룹 스페이스1'),
-       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, null, '그룹 테스트', '그룹 스페이스2'),
-       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, null, '그룹 테스트', '그룹 스페이스3'),
-       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, null, '그룹 테스트', '그룹 스페이스4');
+INSERT INTO workspace (created_at, modified_at, description, division, image_url, subject, title)
+VALUES ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, '개인용도', 'USER1 개인워크스페이스'),
+       ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, '개인용도', 'USER2 개인워크스페이스'),
+       ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, '개인용도', 'USER3 개인워크스페이스'),
+       ('2023-11-11 18:16:21.000000', '2023-11-11 18:16:21.000000', null, 'PERSONAL', null, '개인용도', 'USER4 개인워크스페이스'),
+       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, '그룹 테스트', '그룹 스페이스1'),
+       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, '그룹 테스트', '그룹 스페이스2'),
+       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, '그룹 테스트', '그룹 스페이스3'),
+       ('2023-11-11 18:18:13.000000', '2023-11-11 18:18:13.000000', null, 'GROUP', null, '그룹 테스트', '그룹 스페이스4');
 
 
 insert into workspace_participant(created_at, modified_at, member_id, email, image_url, name, role, workspace_id)
